Function wgpu_subscriber::initialize_default_subscriber

pub fn initialize_default_subscriber(chrome_trace_path: Option<&Path>)

Set up the "standard" logger.

This is fairly inflexible, but a good default to start with. If you need more customization, take what this function does and implement it however you need.

If this function is called, you should not set up a log-based logger like env_logger or fern. This will result in duplicate messages.

Args

chrome_tracing_path if set to Some, will create a trace compatible with chrome://tracing at that location.